pyproject.toml
[project]
name = "latest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.2",
    "asgi-lifespan==2.1.0",
    "asyncpg==0.30.0",
    "certifi==2025.6.15",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.7",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.10",
    "pygments==2.19.2",
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "shellingham==1.5.4",
    "typer==0.16.0",
    "uvicorn==0.34.3",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "databases>=0.9.0",
    "fastapi>=0.115.14",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.41",
]
